auth using next-auth

+++ app/api/staff/route.js
@@ -1,14 +1,22 @@
 import { connectToDB } from "@/utils/database"
 import Staff from "@/models/Staff"
 import { NextResponse } from "next/server"
-import { getDataFromToken } from "@helper/getDataFromToken"
+import { getServerSession } from "next-auth";
+import { options } from "../auth/[...nextauth]/options.js";
 
 export const GET = async (req, res) => {
 
     const { searchParams } = new URL(req.url)
     const id = searchParams.get('staffID') 
-    const userID = await getDataFromToken(req)
     try {
+
+        const session = await getServerSession(options);
+
+        if(!session) {
+            return NextResponse.json({ 'error': "unauthorized"}, { status: 401 })
+        }
+        const userID = session.user.id
+
         await connectToDB();
 
         if(id) {
@@ -27,7 +35,13 @@ export const GET = async (req, res) => {
 export const POST = async (req, res) => {
     const { name, phone, salary } = await req.json()
     try {
-        const userID = await getDataFromToken(req)
+        const session = await getServerSession(options);
+
+        if(!session) {
+            return NextResponse.json({ 'error': "unauthorized"}, { status: 401 })
+        }
+        const userID = session.user.id
+
         await connectToDB()
 
         const newStaff = new Staff({
@@ -49,10 +63,14 @@ export const PUT = async (req, res) => {
     const id = searchParams.get('staffID')
 
     const { name, phone, salary } = await req.json()
-    const userID = await getDataFromToken(req)
 
-    // console.log(staffUpdate)
     try {
+        const session = await getServerSession(options);
+
+        if(!session) {
+            return NextResponse.json({ 'error': "unauthorized"}, { status: 401 })
+        }
+        const userID = session.user.id
         await connectToDB()
 
         const updatedStaff = await Staff.findByIdAndUpdate(id, {
@@ -75,6 +93,12 @@ export const DELETE = async (req, res) => {
     const id = searchParams.get('staffID') 
 
     try {
+        const session = await getServerSession(options);
+
+        if(!session) {
+            return NextResponse.json({ 'error': "unauthorized"}, { status: 401 })
+        }
+        const userID = session.user.id
         await connectToDB()
 
         const deletedStaff = await Staff.findByIdAndDelete(id)
